Обновление запросов оборудования.

+++ src/Application/Events.Application.Services/Features/Equipment/Commands/Delete/DeleteEquipmentHandler.cs
@@ -1,5 +1,7 @@
 ﻿using Events.Application.Services.Shared;
 using Events.Domain.Aggregates.Equipment;
+using Events.Domain.Aggregates.Equipment.Errors;
+using Events.Domain.Exceptions;
 using MediatR;
 
 namespace Events.Application.Services.Features.Equipment.Commands.Delete;
@@ -10,7 +12,11 @@ public sealed class DeleteEquipmentHandler(
 {
     public async Task Handle(DeleteEquipmentCommand request, CancellationToken cancellationToken)
     {
-        var equipment = await equipmentRepository.GetByIdAsync(request.Id, cancellationToken);
+        var equipment = await equipmentRepository.GetByIdAsync(request.EquipmentItemId, cancellationToken);
+
+        if (equipment == null)
+            throw new NotFoundException(EquipmentErrorMessages.NotFoundById(request.EquipmentItemId));
+
         await equipmentRepository.DeleteAsync(equipment, cancellationToken);
     }
 }
